@@ -1,4 +1,4 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -25,5 +25,8 @@ let package = Package(
             name: "SwiftDilithiumTests",
             dependencies: ["SwiftDilithium"],
             resources: [.copy("Resources/kat2.rsp"), .copy("Resources/kat3.rsp"), .copy("Resources/kat5.rsp")]),
+        .testTarget(
+            name: "ForkTests",
+            dependencies: ["SwiftDilithium"]),
     ]
 )
